# gym-track-angular/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/workout-session": {
      "get": {
        "tags": [
          "workout-session-controller"
        ],
        "operationId": "findAllSessions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutSessionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workout-session-controller"
        ],
        "operationId": "saveWorkoutSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/workout-logs": {
      "post": {
        "tags": [
          "workout-logs-controller"
        ],
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutLogsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/plan-exercise/workout/{id}": {
      "post": {
        "tags": [
          "plan-exercise-controller"
        ],
        "operationId": "savePlanExerciseForSpecificWorkout",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanExerciseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/workout-session/{id}": {
      "get": {
        "tags": [
          "workout-session-controller"
        ],
        "operationId": "findSessionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutSessionResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workout-session-controller"
        ],
        "operationId": "deleteWorkoutSession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "workout-session-controller"
        ],
        "operationId": "endSession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/workout-logs/{id}": {
      "get": {
        "tags": [
          "workout-logs-controller"
        ],
        "operationId": "findAllLogsBySessionId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/WorkoutLogsResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workout-logs/records/{id}": {
      "get": {
        "tags": [
          "workout-logs-controller"
        ],
        "operationId": "findAllRecordByExerciseId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutLogsResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workout-logs/plan-exercise/{plan-exercise-id}/session/{session-id}": {
      "get": {
        "tags": [
          "workout-logs-controller"
        ],
        "operationId": "findAllLogsBySessionIdAndPlanExerciseId",
        "parameters": [
          {
            "name": "plan-exercise-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "session-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutLogsResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workout-logs/exercise/{id}": {
      "get": {
        "tags": [
          "workout-logs-controller"
        ],
        "operationId": "findAllLogsByExerciseIdAndUserId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutLogsResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workout-logs/exercise/{exercise-id}/session/{session-id}": {
      "get": {
        "tags": [
          "workout-logs-controller"
        ],
        "operationId": "findAllLogsByExerciseIdAndUserIdGroupedBySessionId",
        "parameters": [
          {
            "name": "exercise-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "session-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/WorkoutLogsResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workout-logs/exercise/max/{id}": {
      "get": {
        "tags": [
          "workout-logs-controller"
        ],
        "operationId": "findAllLogsMaxByExerciseIdAndUserId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutLogsResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/exercise/{exercise_id}": {
      "get": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "findExerciseById",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/exercise/{id}": {
      "delete": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "deletePlanExercise_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "WorkoutLogsRequest": {
        "required": [
          "planExerciseId",
          "reps",
          "setNumber",
          "weight",
          "workoutSessionId"
        ],
        "type": "object",
        "properties": {
          "setNumber": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "maximum": 700,
            "exclusiveMaximum": false,
            "type": "number",
            "format": "double"
          },
          "reps": {
            "type": "integer",
            "format": "int32"
          },
          "workoutSessionId": {
            "type": "integer",
            "format": "int64"
          },
          "planExerciseId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PlanExerciseRequest": {
        "required": [
          "exerciseId"
        ],
        "type": "object",
        "properties": {
          "sets": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "WorkoutSessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "planId": {
            "type": "integer",
            "format": "int64"
          },
          "planName": {
            "type": "string"
          },
          "sessionDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "WorkoutLogsResponse": {
        "type": "object",
        "properties": {
          "setNumber": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "number",
            "format": "double"
          },
          "reps": {
            "type": "integer",
            "format": "int32"
          },
          "summaryWeight": {
            "type": "number",
            "format": "double"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "sessionId": {
            "type": "integer",
            "format": "int64"
          },
          "sessionDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "ExerciseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "CHEST",
              "BACK",
              "BICEPS",
              "SHOULDERS",
              "TRICEPS",
              "FOREARM",
              "ABS",
              "GLUTES",
              "CALF",
              "THIGH",
              "HAMSTRINGS"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}